NASA Temperature Anomalies Data
Global_Temperature_Data_File.txt from NASA and the NOAA represents the change in global surface temperature relative to 1951-1980 average temperatures” and includes both land and ocean temperatures. ​
This code computes the average land temperature and average ocean temperature. The code also calculates the ​variance​ of the values in the file for each column.